Reactivated and enhanced change notifications per issue #11.
This change includes two major components.

First, it reactivates change notifications. That is, Seiso now pushes an
AMQP notification out to RabbitMQ when somebody creates, updates or
deletes an item. (It used to do this, but it's been turned off for a
while.)

The other change is that the design is much cleaner. We use AOP to advise
the ItemService components with notification publication, allowing us to
avoid polluting service code with messaging code. Also, messages now use
the same HAL media type here that the REST API uses.

Note that I expect the details of the representation to change a bit in
response to feedback from some of my Expedia colleagues, so please keep
that in mind should you decide to start consuming notifications.

Reported-by: Willie Wheeler
Signed-off-by: Willie Wheeler