Added v2 support for PUT and DELETE per issue #28
This commit adds support for

* PUTting a single item;
  e.g., PUT https://seiso.example.com/v2/people/jsmith
* PUTting an item property;
  e.g., PUT https://seiso.example.com/v2/people/jsmith/manager
  (include a single URI in the request body, or null to null out the item
  property
* DELETE-ing a single item;
  e.g., DELETE https://seiso.example.com/v2/people/jsmith

In v1 we establish relationships using known fields embedded in the PUT
item request. In v2 we use a hypermedia approach, where we treat the
relationships as resources in their own right, and use URIs (instead of
known fields) to pick out the desired relatum. Hence the introduction of a
new `mergeAssociations` parameter on the `BasicItemDelegate`,
`ItemService`, `ItemMerger` and `ItemSaver`, which allows us to merge
associations selectively: basically for v1 requests, or else v2 PUT item
property requests.