v1 API now returns v1 (not v2) pagination links per issue #39
This change involved a surprising number of files. That's because it
involved refactoring the way we support links for multiple API versions.

Originally the Resource class had v1Links and v2Links fields. The
Resources and PagedResources had only a links field, but on this model
they would have needed v1Links and v2Links as well.

This wasn't a very clean model because it would require modifying existing
classes every time we add or remove an API version, in violation of the
open/closed principle.

Under the new approach, the various Resource classes have a single links
field, but API requests now include a version parameter that supports the
ability to apply different link-building strategies, among other things.
We can also use the approach to effect localized differences between API
versions, such as schema changes, though currently we're using the
serializers to do that. (E.g. HAL serializer suppresses the id field. But
with the API version we could now remove the ID field from the Resource
itself during assembly, which is probably the better approach.)

An alternative approach would be to select entirely different
ResourceAssembly strategies based on the API version, and as APIs diverge
this might become more compelling. This could potentially avoid poluting
the API methods with apiVersion parameters, which occur throughout now,
but this approach has its own challenges, such as somewhat brittle name-
based dependency injection.