Added sourceUri attribute to each importable item type per issue #15
This was mostly some SQL additions (column + index), but it also involved
adding a sourceUri field to the AbstractItem.

The column is nullable, for a couple reasons. First, it will make it
slightly easier to migrate the data over to the new schema. Second, some
entities (like endpoints) won't have associated sources. This may become
a required field in the future.