Changed default Accept header to application/json per issue #48
I had it set to application/hal+json for the v2 API, but Mike Nash noted
that this is a breaking change, and in fact broke his Eos client since he
wasn't setting the Accept header explicitly. So at least for now the
default is back to application/json.

Anyway in the future it might be sensible not to have a default at all--
just require an explicit header.