Fixed data corruption that occurs when moving nodes around per #33.
As the issue notes, the problem was that moving a node from one service
instance to another resulted in a given endpoint having an inconsistent
<nodeIpAddress, serviceInstancePort> pair: they pointed at different
service instances.

The core fix was to delete the node, thereby deleting its dependencies,
and then to re-create the node, thereby recreating its dependencies. This
leaves the IP address roles and service instance ports alone (after all
these exist independently of the node), but it does recreate an endpoint
based on the proper IP address and port.

Also "fixed" the ItemServiceImpl.saveAll() implementation so that it
uses a TransactionTemplate to call ItemServiceImpl.save(). Though I wasn't
observing a problem here before, it seems that this call wouldn't happen
in a transaction, since saveAll() has Propagation.NEVER, and the call to
save() happens behind the transactional proxy.