Fixed GET/PUT /v1/endpoints/{id} per issue #54
The fix involves special treatment for Endpoints, NodeIpAddresses and
ServiceInstancePorts. I don't like this at all, but it stems from the fact
that the v1 API isn't a proper hypermedia API, and in particular uses
ItemKeys and database identifiers to identify items instead of using URIs
as a proper hypermedia API ought to do.

I created issue #55 to address this issue.