Modified global search query to limit size of result set per #40.
We limit the search to 50 items per type. This keeps things fast without
placing an arbitrary limit on the keywords themselves.

The original approach was to require three characters, but if somebody
types in part of a domain name, we could get a huge result set. This is a
problem even with longer keywords. So that's why we're allowing arbitrary
keywords and limiting the result set instead.