Seiso/Seyren integration per #65
This commit adds API and UI support for Seiso/Seyren integration. It
intentionally limits the integration to a many-many relationship between
service instances on the one hand and Seyren checks on the other. This is
limited in two respects.

First, we're not pulling in subscription data. Seyren itself provides a
JSON API, so we simply link to that for now, just to keep things simple,
and also to highlight the hypermedia concept. In the future there might
be some reason to pull this data in (e.g., maybe we want to support
cross-cutting queries related to who get which alerts via which channel),
but there's no such need right now.

Second, Seyren check targets definite an implicit relationship between
checks and machines, but we're attaching checks only to service instances.
Again we're keeping things simple. In the near future we may well apply
a pattern match to bind checks to machines, and then use that relationship
to induce relationships between checks on the one hand, and nodes and
service instances on the other. We're thinking that this may be a nice
discovery mechanism.