Added per-repo find-by-source endpoints per #24
Didn't add anything for User or Role since these aren't items. They still
have the source_id column though, which we may or may not be able to use.
(Maybe we turn these into items after all, or maybe we just attach a
source even where non-items are involved.)

Currently the source_id is not a required column, because none of the
importers support it yet. But we need to update the importers and then
make this column required. Even item in Seiso should have an associated
source.