Added support for linking dashboards to service instances per #68.
This is similar to the Seyren issue (#65), but slightly different in a
couple respects:

1. We're treating dashboards generically, at least for now, as opposed to
   modeling specific vendor technologies.
2. Dashboards can more obviously be associated with entities outside of
   service instances, such as services, teams and so forth.

Due to #2, the service instance owns the relationship between service
instances and dashboards. I ended up moving the service instance/Seyren
check relationship to the service instance as well, just because it seemed
strange to have the service instance own one but not the other. The worry
though is that this won't be a very good model as we add more and more
vendor integrations. (We wouldn't want ten different vendors attached to
the service instance entity.)

So we may end up revisiting that, but for now that's the rationale.