Significant JS cleanup anticipating a series of UI-related tasks.
The main thrust of this admittedly large changeset was to impose some
organization on the AngularJS files, which were previously pretty
haphazard owing to my JS noobishness. Anyway I did a little reading and
thinking to deal with that. :-)

The JS files are now organized roughly around Angular modules. Constants,
controllers, directives, filters and services all appear in their own
files.

Moreover I separated component definition from component registration. For
instance, controller definitions (using array-style annotation to support
eventual minification) appear at the top of ng-controllers.js, and the
registrations appear at the bottom. Easier to read and better separation
of concerns. (Still need to do this to ng-services.js.)

I'm playing around with the page/data load scheme, but it's not where I
want it to be yet.