Added ApiVersion key to projections.
This change will support certain data decouplings we want to do in the v2
API; see #76 and #75. The v1 projections remain coupled for backward
compatibility.

For example, in the v1 API, the single data center projection eagerly
loads the service instances. The single environment projection does the
same. But as it happens, data centers and environments often have hundreds
of service instances, and we don't want to load them all at once, partly
for demand management reasons and partly for UX reasons (pagination).