Prep toward decoupling resource loads from associated collections.
This prep, which continues work from the previous commit, and supports

The service instance details page now loads the service instance and its
nodes separately. We're still not paging the nodes.

Also the UI takes a better approach (not based on JavaScript alert()) to
displaying load errors in the case of failing to load a service instance's
nodes. Will apply this approach elsewhere too in subsequent commits.