Decoupled service instance/meta/node loads per #51
This change makes the service instance details more responsive to the user
and less demanding to the API. Formerly there was a single large query to
grab the service instance info, the node stats and the actual node list.
Now those are three separate calls, and the node list uses pagination.

The work here provides a more general template for how to handle this in
other parts of the app. We aren't currently showing stats elsewhere, but
there are cases where we want to load a resource separately from a large
associated set (e.g., see #75 and #76).

Mike pointed out that we will need a reasonable way to highly problem
nodes, as opposed to forcing people to page through all the nodes. I agree
with that.