Small refactor to use FP idioms and HTTP calls via async.js.
Currently Seiso's JavaScript is written in an imperative style, owing
largely my inexperience with functional programming idioms. This commit
includes two changes of note.

First, in homeController, I replaced the for loops with forEach() calls,
which is more in the functional paradigm. Not sure that what I did is
idiomatic yet, so happy to take suggestions, but the thought is that we
migrate more of the code in this direction.

Second, again in homeController, I'm now using async.map() to effect a
fork/join for the calls to the API. Previousy I was making the calls in
parallel, but wasn't joining, and so occasionally the service groups
would be empty when the call to /v1/services completed first.