Refactor: separate general REST framework code from Seiso-specific code.
At some future point I would like to replace the framework code with
Spring Data REST as it would allow us to get rid of a lot of code. At
this point it doesn't appear to be feasible due to certain seeming bugs
(e.g. https://jira.spring.io/browse/DATAREST-520) but it's something I'd
hope to do.

In preparation for that, and also just to promote better a cleaner arch,
I've created a com.expedia.rf package (rf = "rest framework") intended
for *Seiso-agnostic* framework code. Nothing in this package or its
subpackages should know about the Seiso domain at all. The concept is
that when it's time to switch we should be able to throw com.expedia.rf
away and graft in Spring Data REST.