Added support for display name per issue #147
Chris Brown and I collaborated on this one. This involved three changes:

1) New database column
2) New entity field
3) Front-end updates more or less across the board

There were a few cases where we didn't make the change on the front-end.
These were on the person list (which uses lastName, firstName) and the
search results. The reason for the former is that it would involve parsing
the displayName, and we didn't want to do that yet. The reason for the
latter is that we don't want to confuse people looking for a specific
first name. We can revisit both of these as nobody feels strongly about
them.