Upgraded RabbitMQ integration to use Spring Boot AMQP starter
The Spring Boot AMQP starter handles a lot of the low-level integration
details for us, so we want to take advantage of that.

Using a NotificationGateway abstraction to decouple Seiso from specific
notification mechanisms (like RabbitMQ).

Also fixed a bug in the listeners--we need to send the notifications
*after* the persistence operation, not before, because the notifications
generally trigger client callbacks for Seiso data.