Added node health status reason
This was for issue #215.

Also, as part of this change, I removed the annotated health state
controller and associated components. The reason is that for right now Eos
is using the /nodes/:id and /nodes/:id/healthStatus endpoints, and we need
to think about how we want to handle what are now three closely related
fields. We may end up going with something like a special controller and
service, but I'm disinclined to do that, as I'd rather use existing
endpoints with JSR-303 validators. In any event I'd like to avoid a
commitment until we are actually ready to use these endpoints.