Updated person/group stuff
It turns out that we already had database tables for person_group,
group_membership and group_owner from a couple years back. So we don't
need or even want a new membership table. Moreover, we don't at this
time want membership to be a standalone entity (having a role) because
we're trying to model AD distribution groups, which have no such concept.